+++ mvel/src/commonMain/kotlin/org/islandoftex/arara/mvel/interpreter/AraraExceptionWithHeader.kt
@@ -10,4 +10,5 @@ import org.islandoftex.arara.api.AraraException
 class AraraExceptionWithHeader : AraraException {
     constructor(message: String) : super(message)
     constructor(message: String, exception: Exception) : super(message, exception)
+    constructor(message: String, throwable: Throwable) : super(message, throwable)
 }